+++ dokumentaatio/changelog.md
@@ -1,3 +1,5 @@
 ## Viikko 3
 
 - Lisätty Voucher-luokka
+- Lisätty tietokantayhteys
+- Lisätty CostCentre-luokka